# Speed improvements with -s|--skip