Benchmarks with "time ./pp64.bin -l 10 -s **** < rockyou-top50k.txt"

New skip 100,000,000: 0.095s
Current skip 100,000,000: 2.535s

New skip 1,000,000,000: 0.115s
Current skip 1,000,000,000: 24.583s

New skip 100,000,000,000: 10.520s

I'm still a little disappointed in the new version's speed. There is one more thing that can be done. Initializing db_entries[*]->elems_pos and db_entries[*]->elems_buf[db_entries[*]->elems_pos]->ks_pos before the main loop. I have some ideas as to doing this semi-efficiently. I'll probably do that tonight.