@jsteube 2015-01-12
Btw, the output performance is increased by that patch, too

old:

```
$ time ./pp64.bin -l 100000000 -o /dev/null < rockyou.txt 

real    0m13.033s
user    0m12.884s
sys     0m0.148s
```

new:

```
$ time ./pp64.bin -l 100000000 -o /dev/null < rockyou.txt 

real    0m12.109s
user    0m11.924s
sys     0m0.184s
```

Great work!