@magnumripper 2015-01-12
Kudos @Sc00bz 

JtR uses "skip" when resuming so this was most welcome.

Here are heavier examples: Before:
```
$ time pp64 <rockyou.lst --skip $((1<<32)) | head -1
ehd8694

real    1m2.629s
user    1m1.803s
sys     0m0.811s


$ time pp64 <rockyou.lst --skip $((1<<33)) | head -1
kgitbB

real    2m0.408s
user    1m59.581s
sys     0m0.821s
```
The above indicates that 2^34 would take 4 minutes and 2^37 would take half an hour. I do not even want to calculate how long 2^64 would take...

But here's the new version:
```
$ time pp64 <rockyou.lst --skip $((1<<33)) | head -1
9313oc.s

real    0m2.074s
user    0m1.916s
sys     0m0.156s

$ time pp64 <rockyou.lst --skip $((1<<62)) | head -1
faLLk27sfhso

real    0m2.066s
user    0m1.903s
sys     0m0.163s

$ time pp64 <rockyou.lst --skip `bc <<< 2^64` | head -1
cgtqVONDnbmo

real    0m2.070s
user    0m1.911s
sys     0m0.160s
```
It seems to be totally immune against large numbers :smile: