@Sc00bz 2015-01-13
Oh nice now it's fast. Also that was Atom: https://github.com/jsteube/princeprocessor/commit/2b1082d3e42ebecefae2f95d986f162db0b86c3d