@magnumripper 2015-01-13
Yeah kudos to both of you of course. After the latest bug fix it's not **that** immune to large figures but it's still fast.
```
$ time pp64 <rockyou.lst --skip `bc <<< 2^40` | head -1
Yoylmt7595

real    0m3.413s
user    0m3.251s
sys     0m0.163s

$ time pp64 <rockyou.lst --skip `bc <<< 2^41` | head -1
7QB8y4s

real    0m4.895s
user    0m4.719s
sys     0m0.175s

$ time pp64 <rockyou.lst --skip `bc <<< 2^42` | head -1
rnLP589720

real    0m8.012s
user    0m7.844s
sys     0m0.169s
```