@magnumripper 2015-01-13
Here's that last figure running without GMP though (using gcc's __uint128_t)
```
$ time pp64_u <rockyou.lst --skip `bc <<< 2^42` | head -1
rnLP589720

real    0m3.157s
user    0m2.996s
sys     0m0.160s
```
More than twice as fast.