@Sc00bz 2015-01-14
This is now fixed in https://github.com/jsteube/princeprocessor/pull/12