@magnumripper 2015-01-14
:+1:
```
$ pp64 <rockyou.lst --keyspace 
1584345389899053977143859

$ time pp64 <rockyou.lst --skip 1584345389899053977143850
R...etniq.etniq.
_...etniq.etniq.
$...etniq.etniq.
!...etniq.etniq.
?...etniq.etniq.
i...etniq.etniq.
q...etniq.etniq.
+...etniq.etniq.
....etniq.etniq.

real    0m2.247s
user    0m2.079s
sys     0m0.163s
```